一个默认参数引发的血案。。。。找了两天 model rand = false, 应该为true

+++ c++/utils/matrix/matrix.h
@@ -22,8 +22,8 @@ struct Shape {
     }
 };
 
-typedef double DATATYPE;
-// typedef float DATATYPE;
+// typedef double DATATYPE;
+typedef float DATATYPE;
 
 class Matrix {
